requirements.txt
Django==2.2.2
djangorestframework==3.9.4
mysqlclient==1.4.2.post1
django-cors-headers==3.0.2
gunicorn==19.9.0